# Non-spline Bicubic interpolation added