This makes `TabulatedFluidProperties` much faster than before. 

Timing for 1 million samples (density and derivatives wrt pressure and temperature) for an n by n table of data:

| n | Before | After |
|---|---|---|
| 50 | 4.25 s | 0.35 s |
| 100 | 8.12 s | 0.37 s |

That is a really nice speed up!

Now, `TabulatedFluidProperties` is much faster than calculating water properties and CO2 properties using their UserObjects:

For 1 million samples of density and derivatives wrt pressure and temperature

| Fluid | FluidUserObject | TabulatedFluidProperties |
|---|---|---|
|water | 7.2 s | 0.36 s |
| CO2 | 52.2 s | 0.36 s |


Closes #11059