@permcody 2018-03-20
I'd love to have unit tests for this capability. Can I convince you to add a few?