@cpgr 2018-03-21
Hmm, this failed my own app, which I am sure doesn't use it. I've just compiled it against this branch and it passes on my machine.